Tesis repo updated
This is the complete set of files that run.
The objective of this work is to demonstrate the high sensitivity of the results of an algorithmic operation algorithm according to its parameterization. To this end, variations in a global optimization method will be executed and the results analyzed. The aim is to find statistically significant differences in performance and computational delay in finding the optimal set of parameters in different algorithmic trading strategies. Specifically, variations in the acquisition functions of Bayesian optimization will be studied. The work will use information from the cryptocurrency market due to its various convenient attributes, such as continuous operation, liquidity, and availability of information.